+++ CHANGELOG.md
@@ -6,7 +6,8 @@ X-Core - Changelog
 * SCSS Compiler auf Version v0.6.7 aktualisiert
 * X-Core Styles: Extra MForm Tabs Style entfernt
 * X-Core Styles: MForm Checkbox wieder ausgerichtet
-* rexx API: `sortArticles()` hinzugefügt, thx@fietstouring
+* rexx API: `rexx::sortArticles()` hinzugefügt, thx@fietstouring
+* rexx API: `rexx::threeWayComparison()` hinzugefügt, PHP5 Ersatz für Spaceship Operator `<=>` aus PHP 7
 * Readme: Hinweis zu Unterordner-Installationen hinzugefügt
 
 ### Version 1.0.0 - 15. März 2017